% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{biodbMirbase-package}
\alias{biodbMirbase}
\alias{biodbMirbase-package}
\title{biodbMirbase: biodbMirbase, a library for connecting to miRBase mature database}
\description{
The biodbMirbase library is an extension of the biodb framework package, that provides access to miRBase mature database. It allows to retrieve entries by their accession number, and run specific web services. Description: The biodbMirbase library provides access to the miRBase Database, using biodb package framework.
}
\details{
See vignette biodbMirbase:\preformatted{vignette('biodbMirbase', package='biodbMirbase')
}
}
\seealso{
\link{MirbaseMatureConn}.
}
\author{
\strong{Maintainer}: Pierrick Roger \email{pierrick.roger@cea.fr} (\href{https://orcid.org/0000-0001-8177-4873}{ORCID})

}
